+++ pulsar-inttest-lib/src/main/java/org/apache/pulsar/tests/integration/containers/OpenTelemetryCollectorContainer.java
